+++ celescope/flv_CR/refine_vdj.py
@@ -10,8 +10,7 @@ from collections import defaultdict
 from celescope.tools import utils
 from celescope.tools.step import s_common, Step
 from celescope.tools.emptydrop_cr import get_plot_elements
-from celescope.tools.plotly_plot import Bar_plot
-from celescope.flv_CR.match import gen_clonotypes_table, gen_vj_annotation_metrics
+from celescope.flv_CR.match import gen_vj_annotation_metrics
 from jinja2 import Environment, FileSystemLoader, select_autoescape
 
 
@@ -21,126 +20,128 @@ env = Environment(
     )
 
 
+def get_cltypes(df_contig):
+    """
+    有一个pandas dataframe,名称为df_contig, 包含如下column：['barcode','chain','cdr3','cdr3_nt']，进行如下操作：
+    groupby barcode, 生成一个新的dataframe叫做df_chain，包含新的两列：'cdr3_aa'是{chain}:{cdr3},多个值用分号分隔；‘'cdr3_nt’是{chain}:{cdr3_nt},多个值用分号分隔。
+    将df_chain按照cdr3_aa进行groupby, agg生成一个新的dataframe叫做df_cl,barcode这一列进行计数，这一列叫做'frequency'。cdr3_nt这一列如果相同，则保留一个，如果不同，则用分号连接。
+    将df_cl按照frequency降序排列。
+    """
+    grouped = df_contig.groupby('barcode')
+
+    def concatenate_values(x):
+        cdr3_aa = ';'.join(f"{chain}:{cdr3}" for chain, cdr3 in zip(x['chain'], x['cdr3']))
+        cdr3_nt = ';'.join(f"{chain}:{cdr3_nt}" for chain, cdr3_nt in zip(x['chain'], x['cdr3_nt']))
+        return pd.Series({'cdr3s_aa': cdr3_aa, 'cdr3s_nt': cdr3_nt})
+
+    df_chain = grouped.apply(concatenate_values).reset_index()
+    barcode_cdr3s_aa_dic = {}
+    for _, row in df_chain.iterrows():
+        barcode_cdr3s_aa_dic[row['barcode']] = row['cdr3s_aa']
+
+    df_cl = df_chain.groupby('cdr3s_aa').agg({'barcode': 'count', 'cdr3s_nt': lambda x: ';'.join(x.unique())}).reset_index()
+    df_cl.rename(columns={'barcode': 'frequency'},inplace=True)
+    df_cl = df_cl.sort_values('frequency', ascending=False)
+
+    return df_cl, barcode_cdr3s_aa_dic
+
+def merge_cltypes(cl_origin, cl_refine):
+    """
+    有两个pandas dataframe：cl_origin和cl_refine,cl_origin包含column：[clonotype_id,frequency,cdr3s_aa,cdr3s_nt], cl_refine包含column:[frequency,cdr3s_aa,cdr3s_nt]
+    对cl_refine的每一行做如下处理：如果cdr3s_aa在cl_origin的cdr3s_aa中出现，则将这一行的frequency加在cl_origin第一次出现位置的frequency上，然后从cl_refine中去掉这一行
+    将cl_origin和cl_refine合并为df_merge
+    为df_merge新生成一列proportion, 值为每一行frequency除以frequncy的sum，保留六位小数。
+    将df_merge clonotype为NA的行设为"refine_{num}", num从1开始递增
+    """
+    cl_origin_copy = cl_origin.copy()
+    cl_refine_copy = cl_refine.copy()
+
+    # 遍历cl_refine的每一行
+    for index, row in cl_refine_copy.iterrows():
+        cdr3s_aa = row['cdr3s_aa']
+        mask = cl_origin_copy['cdr3s_aa'] == cdr3s_aa
+        
+        # 如果cdr3s_aa在cl_origin的cdr3s_aa中出现，则将frequency相加
+        if mask.any():
+            first_occurrence_index = cl_origin_copy.loc[mask].index[0]
+            cl_origin_copy.loc[first_occurrence_index, 'frequency'] += row['frequency']
+        
+        # 从cl_refine中去掉相应行
+            cl_refine_copy.drop(index, inplace=True)
+
+    df_merge = pd.concat([cl_origin_copy, cl_refine_copy])
+    df_merge.sort_values(['frequency'], ascending=[False], inplace=True)
+
+    # 计算proportion列
+    df_merge['proportion'] = df_merge['frequency'] / df_merge['frequency'].sum()
+    df_merge['proportion'] = df_merge['proportion'].round(6)
+
+
+    # 将clonotype_id为NA的行设为"refine_{num}"
+    na_rows = df_merge['clonotype_id'].isna()
+    df_merge.loc[na_rows, 'clonotype_id'] = df_merge[na_rows].index.to_series().apply(lambda x: f"refine_{x+1}")
+
+    cdr3s_aa_cl_dic = {}
+    for _, row in df_merge.iterrows():
+        cdr3s_aa_cl_dic[row['cdr3s_aa']] = row['clonotype_id']
+
+    return df_merge,cdr3s_aa_cl_dic
+
+
 class Filter_noise:
     """ Filter noise barcode.
-    auto:
-    keep paired-chain barcode when highest umi of contig >= coeff * second highest umi of contig.
-    
-    snr:
-    https://www.nature.com/articles/s41598-017-18303-z#:~:text=The%20signal%2Dto%2Dnoise%20ratio,of%20background%20pixel%20intensity%2C%20respectively.
-    keep paired-chain barcode when SNR value >= coefficient.
-    The signal-to-noise ratio (SNR) of the single molecules was defined as SNR = (S − B)/σ,
-    where S is the peak single molecule pixel intensity and B and 
-    σ are the average and standard deviation of background pixel intensity, respectively.
-    
-    not_filter:
-    for paired-chain barcode, keep highest tra and trb chain.
+    For is_cell=False barcodes, keep the highest UMI chain when its UMI >= coeff * second largest UMI. Then only paired-chain barcodes are keeped.
+
+    First sort in descending order according to barcode, chain, umis, so that the next step does not need to be sorted Perform groupby by barcode and chain.
+    If each groupby object has only one data, keep it; if it is more than 1, check whether the highest umi is greater than the second highest umi*1.5, if it is satisfied, keep the highest one, if not, then all not reserved Perform groupby by barcode. 
+    If the chain number of each barcode is not equal to 2, it will not be retained.
+    When retaining, use a set to record the contig_id, and after all data checks are completed, determine which rows need to be retained according to the contig_id.
     """
     
-    def __init__(self, df, filter_method, coeff, seqtype):
+    def __init__(self, df, coeff, seqtype):
 
         self.df = df.sort_values("umis", ascending=False)
-        self.method = filter_method
         self.coeff = coeff
         self.seqtype = seqtype
     
     @utils.add_log
     def __call__(self):
         
-        if self.method == "not_filter":
-            self.df = self.not_filter(self.df, self.seqtype)
-        
-        else:
-            bc_chain_dict = self.df.groupby("barcode")["chain"].apply(set).to_dict()
-            if self.seqtype == "TCR":
-                pair_chain_dict = {key: value for key, value in bc_chain_dict.items() if len(value)==2}
+        df_sorted = self.df.sort_values(['barcode', 'chain', 'umis'], ascending=[False, False, False])
+
+        # 按barcode和chain进行groupby，保留满足条件的数据
+        grouped = df_sorted.groupby(['barcode', 'chain'])
+        filtered_data = pd.DataFrame(columns=df_sorted.columns)  # 创建一个空的DataFrame来存储过滤后的数据
+        contig_id_set = set()  # 用于记录需要保留的contig_id
+
+        for _group_name, group_data in grouped:
+            if len(group_data) == 1:
+                contig_id_set.add(group_data['contig_id'].values[0])  
             else:
-                pair_chain_dict = {key: value for key, value in bc_chain_dict.items() if len(value)>=2 and "IGH" in value}
-            self.df = self.df[self.df["barcode"].isin(pair_chain_dict)]
-            self.df = self.df.sort_values(["barcode","umis"], ascending=[False, False])
-            
-            # multi heavy or light chains
-            df_multi_chain = self.df.groupby("barcode").filter(lambda x: (len(x) > 2))
-            # single heavy-light pair chain
-            df_pair_chain = self.df.groupby("barcode").filter(lambda x: (len(x) == 2))
-            
-            if self.seqtype == "TCR":
-                df_heavy = df_multi_chain[df_multi_chain['chain'] == 'TRB']
-                df_light = df_multi_chain[df_multi_chain['chain'] == 'TRA']
-            else:
-                df_heavy = df_multi_chain[(df_multi_chain['chain'] == 'IGH')] 
-                df_light = df_multi_chain[(df_multi_chain['chain'] == 'IGL') | (df_multi_chain['chain'] =='IGK')]
-            light_dict = df_light.groupby("barcode")["umis"].apply(lambda x: x.tolist()).to_dict()
-            heavy_dict = df_heavy.groupby("barcode")["umis"].apply(lambda x: x.tolist()).to_dict()
+                sorted_umis = group_data['umis'].values
+                highest_umi = sorted_umis[0]
+                second_highest_umi = sorted_umis[1]
+                if highest_umi > second_highest_umi * self.coeff:
+                    contig_id_set.add(group_data.iloc[0]['contig_id'])  # 记录保留的contig_id
 
-            if self.method == "auto":
-                for chain_dict in [light_dict, heavy_dict]:
-                    for k in chain_dict:
-                        if len(chain_dict[k]) == 1:
-                            chain_dict[k].append(0.1)
+        filtered_data = self.df[self.df['contig_id'].isin(contig_id_set)]
 
-                # highest umi >= coeff * second highest umi
-                filter_noise_light = {key for key,value in light_dict.items() if value[0]/value[1] >= self.coeff}
-                filter_noise_heavy = {key for key,value in heavy_dict.items() if value[0]/value[1] >= self.coeff}
-                
-            elif self.method == "snr":
-                for chain_dict in [light_dict, heavy_dict]:
-                    for k in chain_dict:
-                        if len(chain_dict[k]) == 1:
-                            chain_dict[k].append(chain_dict[k][0])
-                
-                # snr value >= coeff
-                filter_noise_light = self.snr_filter(light_dict, self.coeff)
-                filter_noise_heavy = self.snr_filter(heavy_dict, self.coeff)
-            
-            else:
-                raise ValueError("Invalid filter method")
-                        
-            filter_noise_barcode = filter_noise_light & filter_noise_heavy | set(df_pair_chain.barcode)
-            self.df = self.df[self.df["barcode"].isin(filter_noise_barcode)]
-        
-        return self.df
+        grouped_barcode = filtered_data.groupby('barcode')
+        final_filtered_data = pd.DataFrame(columns=filtered_data.columns)
+        contig_id_set = set()
 
-    @staticmethod
-    def not_filter(df, seqtype):
-        """do not filter any noise, keep all paired-chain barcodes"""
-        
-        if seqtype == "TCR":
-            df = df.groupby(["barcode", "chain"], as_index=False).head(1)
-            df = df.groupby("barcode").filter(lambda x: (len(x) > 1))
-        else:
-            df_chain_heavy = df[(df['chain'] == 'IGH')]
-            df_chain_light = df[(df['chain'] == 'IGL') | (df['chain'] =='IGK')]
-            df_chain_heavy.drop_duplicates(['barcode'], inplace=True)
-            df_chain_light.drop_duplicates(['barcode'], inplace=True)
-            df = pd.concat([df_chain_heavy, df_chain_light])
-        
-        return df
-    
-    @staticmethod
-    def snr_filter(chain_dict, coeff):
-        """ calculate SNR for each barcode
-        SNR = (S − B)/σ
-        
-        :param chain_dict: {'AAACATCGAAGGACACCCTAATCC': [7, 3, 1],
-                         'AAACATCGAATCCGTCCATCAAGT': [10, 2, 1],
-                         'AAACATCGAATCCGTCCCGACAAC': [4, 3],
-                         }
-        :return: {AAACATCGAAGGACACCCTAATCC, AAACATCGAATCCGTCCATCAAGT}
-        """
-        filter_noise_barcode = set()
-        for k, v in chain_dict.items():
-            S, noise = v[0], v[1:]
-            B = np.mean(noise)
-            O = np.std(noise)
-            if O == 0:
-                filter_noise_barcode.add(k)
-                continue
-            SNR = (S - B) / O
-            if SNR >= coeff:
-                filter_noise_barcode.add(k)
+        for _group_name, group_data in grouped_barcode:
+            if len(group_data['chain'].unique()) == 2:
+                if self.seqtype == 'BCR' and 'IGH' not in group_data['chain'].unique():
+                    continue
+                contig_id_set.update(set(group_data['contig_id']))  # 记录保留的contig_id
 
-        return filter_noise_barcode
+        # 根据contig_id确定需要保留的行
+        final_filtered_data = filtered_data[filtered_data['contig_id'].isin(contig_id_set)]
+        final_filtered_data = final_filtered_data.sort_values(['barcode'])
+
+        return final_filtered_data
+
 
 
 class Refine_vdj(Step):
@@ -148,8 +149,6 @@ class Refine_vdj(Step):
     ## Features
     
     - Refine barcodes where "is_cell=False" and have multi productive chains.
-    
-    - There are three methods to filter noise: SNR, AUTO, NOT_FILTER
 
     """
 
@@ -162,9 +161,10 @@ class Refine_vdj(Step):
         self.all_contig_fasta = glob.glob(f"{self.assemble_out}/all_contig.fasta")[0]
         self.filter_contig_fasta = glob.glob(f"{self.assemble_out}/filtered_contig.fasta")[0]
         self.all_bam = glob.glob(f"{self.assemble_out}/all_contig.bam")[0]
+        self.cl_types = f"{self.assemble_out}/clonotypes.csv"
+        self.df_cl_origin = pd.read_csv(self.cl_types)
         
         self.seqtype = args.seqtype
-        self.filter_method = args.filter_method
         self.coeff = float(args.coeff)
         self.match_cell_barcodes = {}
         
@@ -173,6 +173,9 @@ class Refine_vdj(Step):
 
         with open(args.barcode_convert_json, 'r') as f:
             self.tenX_sgr = json.load(f)
+
+        # out files
+        self.out_cl_file = f'{self.outdir}/clonotypes.csv'
 
 
     def run(self):
@@ -185,8 +188,20 @@ class Refine_vdj(Step):
     def refine(self):
         df_productive = self.all_contig_anno[self.all_contig_anno["productive"] == True]
         df_refine = df_productive[df_productive["is_cell"] == False]
-        df_refine = Filter_noise(df_refine, self.filter_method, self.coeff, self.seqtype)()
+        df_refine = df_refine.groupby("barcode").filter(lambda x: (len(x) > 4))
+        df_refine = Filter_noise(df_refine, self.coeff, self.seqtype)()
+
+        # clonotypes
+        df_cl_refine,barcode_cdr3s_aa_dic = get_cltypes(df_refine)
+        df_cl_merge,cdr3s_aa_cl_dic = merge_cltypes(self.df_cl_origin,df_cl_refine)
+        df_cl_merge.to_csv(self.out_cl_file, index=False)
+        bc_cl_dic = {}
+        for bc in barcode_cdr3s_aa_dic:
+            bc_cl_dic[bc] = cdr3s_aa_cl_dic[barcode_cdr3s_aa_dic[bc]]        
         
+        # add clonotypes
+        df_refine['raw_clonotype_id'] = df_refine['barcode'].replace(bc_cl_dic)
+
         df_merge = pd.concat([self.filter_contig_anno, df_refine])
         refine_cells = set(df_merge.barcode)
         refine_contigs = set(df_merge.contig_id)
@@ -295,33 +310,26 @@ class Refine_vdj(Step):
                     if i["name"] == k:
                         i["display"] = f'{round(v / len(set(df_match.barcode)) * 100, 2)}%'
 
-            gen_clonotypes_table(df_match, f"{self.outdir}/matched_clonotypes.csv")
 
         """
         Clonotypes table
         """
-        gen_clonotypes_table(df_merge, f"{self.outdir}/clonotypes.csv")
+        df_cl = pd.read_csv(self.out_cl_file)
         title = 'Clonetypes'
-        raw_clonotypes = pd.read_csv(f"{self.outdir}/clonotypes.csv", sep=',', index_col=None)
-        raw_clonotypes['ClonotypeID'] = raw_clonotypes['clonotype_id'].apply(lambda x: x.strip('clonetype'))
-        raw_clonotypes['Frequency'] = raw_clonotypes['frequency']
-        raw_clonotypes['Proportion'] = raw_clonotypes['proportion'].apply(lambda x: f'{round(x*100, 2)}%')
-        raw_clonotypes['CDR3_aa'] = raw_clonotypes['cdr3s_aa'].apply(lambda x: x.replace(';', '<br>'))
 
         table_dict = self.get_table_dict(
             title=title,
             table_id='clonetypes',
-            df_table=raw_clonotypes[['ClonotypeID', 'CDR3_aa', 'Frequency', 'Proportion']]
+            df_table=df_cl[['clonotype_id', 'cdr3s_aa', 'frequency', 'proportion']]
         )
         data["match_summary"]["table_dict"] = table_dict
 
-        raw_clonotypes['ClonotypeID'] = raw_clonotypes['ClonotypeID'].astype("int")
-        raw_clonotypes.sort_values(by=['ClonotypeID'], inplace=True)
-        Barplot = Bar_plot(df_bar=raw_clonotypes).get_plotly_div()
-        data["match_summary"]["Barplot"] = Barplot
 
+        raw_report_html = f"{self.outdir}/../.{self.sample}_raw_report.html"
+        report_html = f"{self.outdir}/../{self.sample}_report.html"
+        os.system(f"mv {report_html} {raw_report_html}")
+        
         template = env.get_template(f'html/{self.assay}/base.html')
-        report_html = f"{self.outdir}/{self.sample}_refine_report.html"
         with io.open(report_html, 'w', encoding='utf8') as fh:
             html = template.render(data)
             fh.write(html)
@@ -336,12 +344,10 @@ def refine_vdj(args):
 
 
 def get_opts_refine_vdj(parser, sub_program):
-    parser.add_argument("--not_refine", help="Run the VDJ refine step. ", action="store_true")
-    parser.add_argument("--filter_method", help="filter noise method", choices=["auto", "snr", "not_filter"], default="auto")
+    parser.add_argument("--not_refine", help="Do not perform the refine step. ", action="store_true")
     parser.add_argument("--coeff",
-                        help="coefficient will affect auto and snr noise filter, recommend 1.5 for auto, 10 for snr",
-                        default=1.5
-                        )
+        help="coefficient will affect auto and snr noise filter, recommend 1.5 for auto", 
+        type=float, default=1.5)
     parser.add_argument("--seqtype", help="TCR or BCR", choices=["TCR", "BCR"], required=True)
     if sub_program:
         s_common(parser)