+++ celescope/tools/plotly_plot.py
@@ -55,7 +55,7 @@ class Plotly_plot():
     def plotly_plot(self):
         return plotly.offline.plot(
             self._fig,
-            include_plotlyjs=True,
+            include_plotlyjs=False,
             output_type='div',
             config=PLOTLY_CONFIG
         )